split local volumes reducer
close DCOS-14967
