+++ plugins/services/src/js/components/forms/VolumesFormSection.js
@@ -3,7 +3,7 @@ import React, {Component} from 'react';
 import Objektiv from 'objektiv';
 
 import {FormReducer as externalVolumes} from '../../reducers/serviceForm/FormReducers/ExternalVolumes';
-import {FormReducer as localVolumes} from '../../reducers/serviceForm/LocalVolumes';
+import {FormReducer as localVolumes} from '../../reducers/serviceForm/FormReducers/LocalVolumes';
 import AddButton from '../../../../../../src/js/components/form/AddButton';
 import FieldError from '../../../../../../src/js/components/form/FieldError';
 import FieldInput from '../../../../../../src/js/components/form/FieldInput';
